📝 docs: Add MIT License and update README.md 🔖 feat: Add MIT License to the project The MIT License has been added to the project to provide clear guidelines for what other developers can and cannot do with the software. This is a common license for open source projects and allows others to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the software.
📚 docs(README.md): Update README.md with more specific information
The README.md file has been updated to provide more specific information about the project. This includes a link to the creator's GitHub profile and a footnote explaining what default community health files are. This makes it easier for other developers to understand the purpose and guidelines of the project.